rename components using kebab case
