+++ src/components/stream-info.tsx
