+++ src/components/vision-selector.tsx
